app/requirements.txt
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
debtcollector==3.0.0
decorator==5.1.1
dogpile.cache==1.3.3
Flask==3.1.0
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
keystoneauth1==5.9.0
MarkupSafe==3.0.2
msgpack==1.1.0
netaddr==1.3.0
netifaces==0.11.0
openstacksdk==4.1.0
os-service-types==1.7.0
oslo.config==9.7.0
oslo.i18n==6.5.0
oslo.serialization==5.6.0
oslo.utils==8.0.0
packaging==24.2
pbr==6.1.0
platformdirs==4.3.6
psutil==6.1.0
pycparser==2.22
pyparsing==3.2.0
python-keystoneclient==5.5.0
PyYAML==6.0.2
requests==2.32.3
requestsexceptions==1.4.0
rfc3986==2.0.0
schedule==1.2.2
stevedore==5.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
